backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
langgraph==0.2.50
langchain==0.3.7
langchain-openai==0.2.7
langchain-core==0.3.16
langchain-ollama==0.2.0 
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
dnspython==2.7.0
openpyxl==3.1.5
python-dotenv==1.0.1